+++ src/Firebase.js
@@ -1,12 +1,5 @@
 import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore";
-import {
-  getAuth, // authentication 설정
-  signInWithPopup, //google 로그인을 팝업창에 띄우기 위해
-  GoogleAuthProvider, //google login 기능
-  signInWithEmailAndPassword, // email 로그인
-  createUserWithEmailAndPassword, //email 회원가입
-} from "firebase/auth";
 
 const firebaseConfig = {
   apiKey: process.env.REACT_APP_FIREBASE_APIKEY,
@@ -20,6 +13,5 @@ const firebaseConfig = {
 
 const firebase = initializeApp(firebaseConfig);
 const fireStore = getFirestore(firebase);
-const authService = getAuth();
 
-export { fireStore, firebase, authService };
+export { fireStore, firebase };
